requirements.txt
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
isodate==0.6.1
tqdm==4.66.4
openpyxl==3.1.2

## just run " pip install -r requirements.txt " in terminal 